move the logic if we should fail into the GitDirLocator
